Fix inlining issue
